Separaçao dos .cpp e inclusões
